fix: drift filters you can send, and one dialect for dates
The Drift queue's filters were two useState calls, so a narrowed queue had no
link: an operator who scoped to one project and sent it to a colleague sent
them the whole backlog, and a refresh threw the narrowing away. They live in
the URL now, the way People's already do, and drift-filters.ts is the one
place that parses them — so any screen can link into a pre-filtered queue.

Four dimensions added. Person and role are built from the rows actually in the
queue rather than from the whole directory, so nothing on offer returns
nothing. Age asks how long a finding has been sitting there. Origin is the
closest thing to "why can't Syndra explain it", and is deliberately not called
that: a PENDING row has no recorded explanation by construction — if a bundle,
rule or grant explained it, the sweep would never have raised it. The honest
axis is how much is known about who did it, which is what decides whether an
operator can attribute it or must judge it blind. Three of the six are request
parameters and three are applied over what came back; which is which is stated
on each, because a filter that silently meant "of the first page" would be a
worse lie than no filter at all.

On Home: the greeting wraps and stops short of the window edge rather than
running into it; "49 things here need you" loses the word that made it read
oddly; and the card headed "New people without a bundle" no longer counts a
missing default bundle as if it were a person — that setting is a real thing
to do, so the headline still counts it, but a badge on a card about people is
not the place to say so.

And one dialect for dates: the audit log showed a bare "18 Sept" where the
same event in Lately showed "18 Sept 02:57", and Home said "19 September" in
full beside both. All three read the same now.

+++ ui/src/lib/format.ts
@@ -161,10 +161,13 @@ export function formatLongDate(iso: string | null | undefined): string {
 
 /** "Friday 31 July" — the line under a greeting. */
 export function formatWeekday(date: Date = new Date()): string {
+  // Short month, matching formatShortDate everywhere else. "Saturday 19
+  // September" beside a feed of "18 Sept" rows is the same product speaking
+  // two dialects on one screen.
   return date.toLocaleDateString(DATE_LOCALE, {
     weekday: "long",
     day: "numeric",
-    month: "long",
+    month: "short",
   });
 }
